tests: hasRolePermissions tests
